Add sources table
Adds a table to the Muzei ContentProvider for retrieving information about the sources used with Muzei

+++ android-client-common/src/main/java/com/google/android/apps/muzei/provider/MuzeiProvider.java
@@ -17,6 +17,7 @@
 package com.google.android.apps.muzei.provider;
 
 import android.content.ContentProvider;
+import android.content.ContentUris;
 import android.content.ContentValues;
 import android.content.Context;
 import android.content.Intent;
@@ -32,6 +33,7 @@ import android.os.ParcelFileDescriptor;
 import android.preference.PreferenceManager;
 import android.provider.BaseColumns;
 import android.support.annotation.NonNull;
+import android.text.TextUtils;
 import android.util.Log;
 
 import com.google.android.apps.muzei.api.MuzeiContract;
@@ -54,21 +56,35 @@ public class MuzeiProvider extends ContentProvider {
      */
     private static final int ARTWORK = 1;
     /**
+     * The incoming URI matches the SOURCE URI pattern
+     */
+    private static final int SOURCES = 2;
+    /**
+     * The incoming URI matches the SOURCE ID URI pattern
+     */
+    private static final int SOURCE_ID = 3;
+    /**
      * The database that the provider uses as its underlying data store
      */
     private static final String DATABASE_NAME = "muzei.db";
     /**
      * The database version
      */
-    private static final int DATABASE_VERSION = 2;
+    private static final int DATABASE_VERSION = 3;
     /**
      * A UriMatcher instance
      */
     private static final UriMatcher uriMatcher = MuzeiProvider.buildUriMatcher();
     /**
-     * An identity all column projection mapping
+     * An identity all column projection mapping for Artwork
      */
-    private final HashMap<String, String> allColumnProjectionMap = MuzeiProvider.buildAllColumnProjectionMap();
+    private final HashMap<String, String> allArtworkColumnProjectionMap =
+            MuzeiProvider.buildAllArtworkColumnProjectionMap();
+    /**
+     * An identity all column projection mapping for Sources
+     */
+    private final HashMap<String, String> allSourcesColumnProjectionMap =
+            MuzeiProvider.buildAllSourcesColumnProjectionMap();
     /**
      * Handle to a new DatabaseHelper.
      */
@@ -92,11 +108,11 @@ public class MuzeiProvider extends ContentProvider {
     }
 
     /**
-     * Creates and initializes a column project for all columns
+     * Creates and initializes a column project for all columns for Artwork
      *
-     * @return The all column projection map
+     * @return The all column projection map for Artwork
      */
-    private static HashMap<String, String> buildAllColumnProjectionMap() {
+    private static HashMap<String, String> buildAllArtworkColumnProjectionMap() {
         final HashMap<String, String> allColumnProjectionMap = new HashMap<>();
         allColumnProjectionMap.put(BaseColumns._ID, BaseColumns._ID);
         allColumnProjectionMap.put(MuzeiContract.Artwork.COLUMN_NAME_IMAGE_URI,
@@ -117,6 +133,27 @@ public class MuzeiProvider extends ContentProvider {
     }
 
     /**
+     * Creates and initializes a column project for all columns for Sources
+     *
+     * @return The all column projection map for Sources
+     */
+    private static HashMap<String, String> buildAllSourcesColumnProjectionMap() {
+        final HashMap<String, String> allColumnProjectionMap = new HashMap<>();
+        allColumnProjectionMap.put(BaseColumns._ID, BaseColumns._ID);
+        allColumnProjectionMap.put(MuzeiContract.Sources.COLUMN_NAME_COMPONENT_NAME,
+                MuzeiContract.Sources.COLUMN_NAME_COMPONENT_NAME);
+        allColumnProjectionMap.put(MuzeiContract.Sources.COLUMN_NAME_IS_SELECTED,
+                MuzeiContract.Sources.COLUMN_NAME_IS_SELECTED);
+        allColumnProjectionMap.put(MuzeiContract.Sources.COLUMN_NAME_DESCRIPTION,
+                MuzeiContract.Sources.COLUMN_NAME_DESCRIPTION);
+        allColumnProjectionMap.put(MuzeiContract.Sources.COLUMN_NAME_WANTS_NETWORK_AVAILABLE,
+                MuzeiContract.Sources.COLUMN_NAME_WANTS_NETWORK_AVAILABLE);
+        allColumnProjectionMap.put(MuzeiContract.Sources.COLUMN_NAME_COMMANDS,
+                MuzeiContract.Sources.COLUMN_NAME_COMMANDS);
+        return allColumnProjectionMap;
+    }
+
+    /**
      * Creates and initializes the URI matcher
      *
      * @return the URI Matcher
@@ -125,6 +162,10 @@ public class MuzeiProvider extends ContentProvider {
         final UriMatcher matcher = new UriMatcher(UriMatcher.NO_MATCH);
         matcher.addURI(MuzeiContract.AUTHORITY, MuzeiContract.Artwork.TABLE_NAME,
                 MuzeiProvider.ARTWORK);
+        matcher.addURI(MuzeiContract.AUTHORITY, MuzeiContract.Sources.TABLE_NAME,
+                MuzeiProvider.SOURCES);
+        matcher.addURI(MuzeiContract.AUTHORITY, MuzeiContract.Sources.TABLE_NAME + "/#",
+                MuzeiProvider.SOURCE_ID);
         return matcher;
     }
 
@@ -140,8 +181,14 @@ public class MuzeiProvider extends ContentProvider {
          */
         switch (MuzeiProvider.uriMatcher.match(uri)) {
             case ARTWORK:
-                // If the pattern is for artwork, returns the general content type.
+                // If the pattern is for artwork, returns the artwork content type.
                 return MuzeiContract.Artwork.CONTENT_TYPE;
+            case SOURCES:
+                // If the pattern is for sources, returns the sources content type.
+                return MuzeiContract.Sources.CONTENT_TYPE;
+            case SOURCE_ID:
+                // If the pattern is for source id, returns the sources content item type.
+                return MuzeiContract.Sources.CONTENT_ITEM_TYPE;
             default:
                 throw new IllegalArgumentException("Unknown URI " + uri);
         }
@@ -149,9 +196,16 @@ public class MuzeiProvider extends ContentProvider {
 
     @Override
     public Uri insert(@NonNull final Uri uri, final ContentValues values) {
-        // Validates the incoming URI. Only the full provider URI is allowed for inserts.
-        if (MuzeiProvider.uriMatcher.match(uri) != MuzeiProvider.ARTWORK)
+        if (MuzeiProvider.uriMatcher.match(uri) == MuzeiProvider.ARTWORK) {
+            return insertArtwork(uri, values);
+        } else if (MuzeiProvider.uriMatcher.match(uri) == MuzeiProvider.SOURCES) {
+            return insertSource(uri, values);
+        } else {
             throw new IllegalArgumentException("Unknown URI " + uri);
+        }
+    }
+
+    private Uri insertArtwork(@NonNull final Uri uri, final ContentValues values) {
         if (values == null) {
             throw new IllegalArgumentException("Invalid ContentValues: must not be null");
         }
@@ -170,6 +224,29 @@ public class MuzeiProvider extends ContentProvider {
         return MuzeiContract.Artwork.CONTENT_URI;
     }
 
+    private Uri insertSource(@NonNull final Uri uri, final ContentValues initialValues) {
+        ContentValues values;
+        if (initialValues != null)
+            values = new ContentValues(initialValues);
+        else
+            values = new ContentValues();
+        if (!values.containsKey(MuzeiContract.Sources.COLUMN_NAME_COMPONENT_NAME))
+            throw new IllegalArgumentException("Initial values must contain component name " + initialValues);
+        final SQLiteDatabase db = databaseHelper.getWritableDatabase();
+        final long rowId = db.insert(MuzeiContract.Sources.TABLE_NAME,
+                MuzeiContract.Sources.COLUMN_NAME_COMPONENT_NAME, values);
+        // If the insert succeeded, the row ID exists.
+        if (rowId > 0)
+        {
+            // Creates a URI with the source ID pattern and the new row ID appended to it.
+            final Uri sourceUri = ContentUris.withAppendedId(MuzeiContract.Sources.CONTENT_URI, rowId);
+            getContext().getContentResolver().notifyChange(sourceUri, null);
+            return sourceUri;
+        }
+        // If the insert didn't succeed, then the rowID is <= 0
+        throw new SQLException("Failed to insert row into " + uri);
+    }
+
     /**
      * Creates the underlying DatabaseHelper
      *
@@ -184,14 +261,44 @@ public class MuzeiProvider extends ContentProvider {
     @Override
     public Cursor query(@NonNull final Uri uri, final String[] projection, final String selection,
                         final String[] selectionArgs, final String sortOrder) {
-        // Validates the incoming URI. Only the full provider URI is allowed for queries.
-        if (MuzeiProvider.uriMatcher.match(uri) != MuzeiProvider.ARTWORK)
+        if (MuzeiProvider.uriMatcher.match(uri) == MuzeiProvider.ARTWORK) {
+            return queryArtwork(uri, projection, selection, selectionArgs, sortOrder);
+        } else if (MuzeiProvider.uriMatcher.match(uri) == MuzeiProvider.SOURCES ||
+                MuzeiProvider.uriMatcher.match(uri) == MuzeiProvider.SOURCE_ID) {
+            return querySource(uri, projection, selection, selectionArgs, sortOrder);
+        } else {
             throw new IllegalArgumentException("Unknown URI " + uri);
+        }
+    }
+
+    private Cursor queryArtwork(@NonNull final Uri uri, final String[] projection, final String selection,
+                        final String[] selectionArgs, final String sortOrder) {
         final SQLiteQueryBuilder qb = new SQLiteQueryBuilder();
         qb.setTables(MuzeiContract.Artwork.TABLE_NAME);
-        qb.setProjectionMap(allColumnProjectionMap);
+        qb.setProjectionMap(allArtworkColumnProjectionMap);
         final SQLiteDatabase db = databaseHelper.getReadableDatabase();
         final Cursor c = qb.query(db, projection, selection, selectionArgs, null, null, sortOrder, null);
+        c.setNotificationUri(getContext().getContentResolver(), uri);
+        return c;
+    }
+
+    private Cursor querySource(@NonNull final Uri uri, final String[] projection, final String selection,
+                                final String[] selectionArgs, final String sortOrder) {
+        final SQLiteQueryBuilder qb = new SQLiteQueryBuilder();
+        qb.setTables(MuzeiContract.Sources.TABLE_NAME);
+        qb.setProjectionMap(allSourcesColumnProjectionMap);
+        final SQLiteDatabase db = databaseHelper.getReadableDatabase();
+        if (MuzeiProvider.uriMatcher.match(uri) == SOURCE_ID) {
+            // If the incoming URI is for a single source identified by its ID, appends "_ID = <sourceId>"
+            // to the where clause, so that it selects that single ingredient
+            qb.appendWhere(BaseColumns._ID + "=" + uri.getPathSegments().get(1));
+        }
+        String orderBy;
+        if (TextUtils.isEmpty(sortOrder))
+            orderBy = MuzeiContract.Sources.DEFAULT_SORT_ORDER;
+        else
+            orderBy = sortOrder;
+        final Cursor c = qb.query(db, projection, selection, selectionArgs, null, null, orderBy, null);
         c.setNotificationUri(getContext().getContentResolver(), uri);
         return c;
     }
@@ -219,7 +326,43 @@ public class MuzeiProvider extends ContentProvider {
 
     @Override
     public int update(@NonNull final Uri uri, final ContentValues values, final String selection, final String[] selectionArgs) {
-        throw new UnsupportedOperationException("Updates are not allowed: insert does an insert or update operation");
+        if (MuzeiProvider.uriMatcher.match(uri) == MuzeiProvider.ARTWORK) {
+            throw new UnsupportedOperationException("Updates are not allowed: insert does an insert or update operation");
+        } else if (MuzeiProvider.uriMatcher.match(uri) == MuzeiProvider.SOURCES ||
+                MuzeiProvider.uriMatcher.match(uri) == MuzeiProvider.SOURCE_ID) {
+            return updateSource(uri, values, selection, selectionArgs);
+        } else {
+            throw new IllegalArgumentException("Unknown URI " + uri);
+        }
+    }
+
+    private int updateSource(@NonNull final Uri uri, final ContentValues values, final String selection,
+                             final String[] selectionArgs) {
+        final SQLiteDatabase db = databaseHelper.getWritableDatabase();
+        int count = 0;
+        switch (MuzeiProvider.uriMatcher.match(uri))
+        {
+            case SOURCES:
+                // If the incoming URI matches the general sources pattern, does the update based on the incoming
+                // data.
+                count = db.update(MuzeiContract.Sources.TABLE_NAME, values, selection, selectionArgs);
+                break;
+            case SOURCE_ID:
+                // If the incoming URI matches a single source ID, does the update based on the incoming data, but
+                // modifies the where clause to restrict it to the particular source ID.
+                String finalWhere = BaseColumns._ID + " = " + uri.getPathSegments().get(1);
+                // If there were additional selection criteria, append them to the final WHERE clause
+                if (selection != null)
+                    finalWhere = finalWhere + " AND " + selection;
+                count = db.update(MuzeiContract.Sources.TABLE_NAME, values, finalWhere, selectionArgs);
+                break;
+            default:
+                throw new IllegalArgumentException("Unknown URI " + uri);
+        }
+        if (count > 0) {
+            getContext().getContentResolver().notifyChange(uri, null);
+        }
+        return count;
     }
 
     /**
@@ -240,24 +383,42 @@ public class MuzeiProvider extends ContentProvider {
          */
         @Override
         public void onCreate(final SQLiteDatabase db) {
-            db.execSQL("CREATE TABLE " + MuzeiContract.Artwork.TABLE_NAME + " (" + BaseColumns._ID
-                    + " INTEGER PRIMARY KEY AUTOINCREMENT," + MuzeiContract.Artwork.COLUMN_NAME_IMAGE_URI
-                    + " TEXT," + MuzeiContract.Artwork.COLUMN_NAME_TITLE + " TEXT,"
+            db.execSQL("CREATE TABLE " + MuzeiContract.Artwork.TABLE_NAME + " ("
+                    + BaseColumns._ID + " INTEGER PRIMARY KEY AUTOINCREMENT,"
+                    + MuzeiContract.Artwork.COLUMN_NAME_IMAGE_URI + " TEXT,"
+                    + MuzeiContract.Artwork.COLUMN_NAME_TITLE + " TEXT,"
                     + MuzeiContract.Artwork.COLUMN_NAME_BYLINE + " TEXT,"
                     + MuzeiContract.Artwork.COLUMN_NAME_ATTRIBUTION + " TEXT,"
                     + MuzeiContract.Artwork.COLUMN_NAME_TOKEN + " TEXT,"
                     + MuzeiContract.Artwork.COLUMN_NAME_META_FONT + " TEXT,"
                     + MuzeiContract.Artwork.COLUMN_NAME_VIEW_INTENT + " TEXT);");
+            onCreateSourcesTable(db);
+        }
+
+        private void onCreateSourcesTable(final SQLiteDatabase db) {
+            db.execSQL("CREATE TABLE " + MuzeiContract.Artwork.TABLE_NAME + " ("
+                    + BaseColumns._ID + " INTEGER PRIMARY KEY AUTOINCREMENT,"
+                    + MuzeiContract.Sources.COLUMN_NAME_COMPONENT_NAME + " TEXT,"
+                    + MuzeiContract.Sources.COLUMN_NAME_IS_SELECTED + " INTEGER,"
+                    + MuzeiContract.Sources.COLUMN_NAME_DESCRIPTION + " TEXT,"
+                    + MuzeiContract.Sources.COLUMN_NAME_WANTS_NETWORK_AVAILABLE + " INTEGER,"
+                    + MuzeiContract.Sources.COLUMN_NAME_COMMANDS + " TEXT);");
         }
 
         /**
-         * Demonstrates that the provider must consider what happens when the underlying database is changed. Note that
-         * this currently just destroys and recreates the database - should upgrade in place
+         * Upgrades the database in place.
          */
         @Override
         public void onUpgrade(final SQLiteDatabase db, final int oldVersion, final int newVersion) {
-            db.execSQL("DROP TABLE IF EXISTS " + MuzeiContract.Artwork.TABLE_NAME);
-            onCreate(db);
+            if (newVersion == 2) {
+                db.execSQL("DROP TABLE IF EXISTS " + MuzeiContract.Artwork.TABLE_NAME);
+                onCreate(db);
+                return;
+            }
+            if (oldVersion < 3) {
+                // Add the sources table
+                onCreateSourcesTable(db);
+            }
         }
     }
 }
